+++ pycaption/dfxp/base.py
@@ -122,14 +122,21 @@ class DFXPReader(BaseReader):
         return None
 
     def _find_times(self, p_tag):
-        try:
-            start = self._translate_time(p_tag['begin'])
-        except KeyError:
-            raise CaptionReadTimingError(f'Missing begin time on line {p_tag}')
+        begin = p_tag.get('begin')
+        if not begin:
+            raise CaptionReadTimingError(
+                f'Missing begin time on line {p_tag}.')
 
-        try:
+        end = p_tag.get('end')
+        dur = p_tag.get('dur')
+        if not end and not dur:
+            raise CaptionReadTimingError(
+                f'Missing end time or duration on line {p_tag}.')
+
+        start = self._translate_time(begin)
+        if end:
             end = self._translate_time(p_tag['end'])
-        except KeyError:
+        else:
             dur = self._translate_time(p_tag['dur'])
             end = start + dur